Flytt NavnOgIdent til passende fil og slett da ubrukt fil
