+++ src/main/kotlin/no/nav/familie/ba/sak/ekstern/restDomene/RestJournalføring.kt
@@ -1,5 +1,8 @@
 package no.nav.familie.ba.sak.ekstern.restDomene
 
+import no.nav.familie.ba.sak.common.FunksjonellFeil
+import no.nav.familie.ba.sak.common.erAlfanummeriskPlussKolon
+import no.nav.familie.ba.sak.common.secureLogger
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.Bruker
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.OppdaterJournalpostRequest
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingKategori
@@ -77,6 +80,22 @@ data class RestJournalføring(
             journalpostTittel?.contains("utvidet") == true -> BehandlingUnderkategori.UTVIDET
             // Defaulter til ordinær inntil videre.
             else -> BehandlingUnderkategori.ORDINÆR
+        }
+    }
+}
+
+data class NavnOgIdent(
+    val navn: String,
+    val id: String,
+) {
+    // Bruker init til å validere personidenten
+    init {
+        if (!id.erAlfanummeriskPlussKolon()) {
+            secureLogger.info("Ugyldig ident: $id")
+            throw FunksjonellFeil(
+                melding = "Ugyldig ident. Se securelog for mer informasjon.",
+                frontendFeilmelding = "Ugyldig ident. Normalt et fødselsnummer eller organisasjonsnummer",
+            )
         }
     }
 }
